remove some mistakes after review
